requirements.txt
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.7
Flask==3.1.0
Flask-Login==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
peewee==3.17.8
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.1.3
